ver1/requirements.txt
streamlit
python-docx
pymupdf
Pillow
requests
numpy
